Permissions - user service, agreement-service done

+++ service-user/src/main/java/by/salary/serviceuser/service/UserAgreementsService.java
@@ -1,6 +1,9 @@
 package by.salary.serviceuser.service;
 
+import by.salary.serviceuser.entities.Permission;
+import by.salary.serviceuser.entities.PermissionsEnum;
 import by.salary.serviceuser.entities.UserAgreement;
+import by.salary.serviceuser.exceptions.NotEnoughtPermissionsException;
 import by.salary.serviceuser.exceptions.UserNotFoundException;
 import by.salary.serviceuser.model.UserAgreementRequestDTO;
 import by.salary.serviceuser.model.UserAgreementResponseDTO;
@@ -27,9 +30,15 @@ public class UserAgreementsService {
     }
 
 
-    public List<UserAgreementResponseDTO> getAllUserAgreements(BigInteger userId) {
+    public List<UserAgreementResponseDTO> getAllUserAgreements(BigInteger userId, String email, List<Permission> permissions) {
         if(!userRepository.existsById(userId)){
             throw new UserNotFoundException("User with id " + userId + " not found", HttpStatus.NOT_FOUND);
+        }
+
+        if(!userId.equals(userRepository.findByUserEmail(email).get().getId())
+            & ! permissions.contains(new Permission(PermissionsEnum.READ_USER_AGREEMENT))
+        ){
+            throw new NotEnoughtPermissionsException("User with id " + userId + " not found", HttpStatus.FORBIDDEN);
         }
 
         List<UserAgreementResponseDTO> userAgreements = new ArrayList<>();
@@ -39,17 +48,30 @@ public class UserAgreementsService {
             }
         });
 
+
         return userAgreements;
     }
 
-    public void deleteUserAgreement(BigInteger agreementId) {
+    public void deleteUserAgreement(BigInteger agreementId, String email, List<Permission> permissions) {
+        if(!permissions.contains(new Permission(PermissionsEnum.DELETE_USER_AGREEMENT))){
+            throw new NotEnoughtPermissionsException("You have not enought permissions to perform this action", HttpStatus.FORBIDDEN);
+        }
+        if(!userRepository.existsByUserEmail(email)){
+            throw new UserNotFoundException("User with email " + email + " not found", HttpStatus.NOT_FOUND);
+        }
         if(!userAgreementsRepository.existsById(agreementId)){
             throw new UserNotFoundException("User agreement with id " + agreementId + " not found", HttpStatus.NOT_FOUND);
         }
         userAgreementsRepository.deleteById(agreementId);
     }
 
-    public void deleteUserAgreements(BigInteger userId) {
+    public void deleteUserAgreements(BigInteger userId, String email, List<Permission> permissions) {
+        if(!permissions.contains(new Permission(PermissionsEnum.DELETE_USER_AGREEMENT))){
+            throw new NotEnoughtPermissionsException("You have not enought permissions to perform this action", HttpStatus.FORBIDDEN);
+        }
+        if(!userRepository.existsByUserEmail(email)){
+            throw new UserNotFoundException("User with email " + email + " not found", HttpStatus.NOT_FOUND);
+        }
         if(!userRepository.existsById(userId)){
             throw new UserNotFoundException("User with id " + userId + " not found", HttpStatus.NOT_FOUND);
         }
@@ -62,10 +84,17 @@ public class UserAgreementsService {
         );
     }
 
-    public UserAgreementResponseDTO createUserAgreement(UserAgreementRequestDTO userAgreementRequestDTO, BigInteger userId) {
+    public UserAgreementResponseDTO createUserAgreement(UserAgreementRequestDTO userAgreementRequestDTO, BigInteger userId, String email, List<Permission> permissions) {
+        if(!userRepository.existsByUserEmail(email)){
+            throw new UserNotFoundException("User with email " + email + " not found", HttpStatus.NOT_FOUND);
+        }
+        if(!permissions.contains(new Permission(PermissionsEnum.ADD_USER_AGREEMENT))){
+            throw new NotEnoughtPermissionsException("You have not enought permissions to perform this action", HttpStatus.FORBIDDEN);
+        }
         if(!userRepository.existsById(userId)){
             throw new UserNotFoundException("User with id " + userId + " not found", HttpStatus.NOT_FOUND);
         }
+
         UserAgreement userAgreement = new UserAgreement(userAgreementRequestDTO, userRepository.findById(userId).get());
         return new UserAgreementResponseDTO(userAgreementsRepository.save(userAgreement));
     }
